rustc: Rename `simd128` wasm feature to `wasm-simd128`
This commit applies feedback from #74372 where the current name of the
wasm simd feature, `simd128`, felt a bit too bland and un-specific. This
commit renames the feature to `wasm-simd128`. This doesn't actually
delete the old `simd128` recognized name just yet because the stdarch
submodule needs to update first, so what this commit actually does is
introduce a new name (unstable like it was before), and then once
stdarch is updated we can delete the previously recognized `simd128` name.

Note that other wasm features (none currently stable) are left as-is to
choose an appropriate name if they're proposed for stabilization.