Add full debug layout for type layout doc
See what you can see for `#[rustc_layout(debug)]` but with rustdoc which
is very useful, rather than writing some stuff you can just browse through
the docs. I think it may be useful to put it in json and have something
process it and do sort of `ncdu` to easily optimize on type sizes.