Remove stability marker arrow.
This was originally introduced in # 51387, to solve a UI problem: in a
list of collapsed methods, it was not clear whether the stability marker
attached to the item above it or the one below it.

That problem has now been solved in a couple of different ways:

 - The margin between the stability marker and the item it annotates is
   much smaller, indicating their relationship.
 - Stability markers are now collapsed along with the docblock.

Also, the arrow made our CSS fragile, since it was positioned absolutely
and needed to line up horizontally with the beginning of the item (or
with the link icon).

Aligning this with the rest of the docblock makes getting the CSS right
a lot easier.