Mark `SocketAncillary::add_fds` unsafe.
The `SocketAncillary` object has no lifetime that could guarantee that
the file descriptor remains open long enough. That means it could end up
duplicating arbitrary file descriptors and send them to a remote
process.

Additionally, even if the fd lifetime was tracked, duplicating an FD and
sending it to a remote process could violate the safety requirements of
the object managing the file descriptor. This matters, because
`as_raw_fd()` is not unsafe.