proc_macro: add support for sending messages without waiting for a reply
This has minimal impact without changes from a later part, as it also requires
support from the ExecutionStrategy. Messages are able to return owning handles
without waiting by allocating an ID from the client thread, and passing it up
to the server to fill with a response. If the server panics, it will be an ICE,
and no further requests will be handled.