rustdoc: Unify external_paths and paths cache map
The `formats::cache::Cache` fields `paths` and `external_paths` are now
unified to a single `paths` map, which makes it easier to manage the
paths and avoids using local paths as extern and vice versa.

The `paths` map maps a `DefId` to an enum which can be either `Local` or
`Extern` so there can't be any misusing.