Optimize `Iterator::map_windows` by using a buffer of size 2 * N
This means one iteration is amortized O(1) instead of O(N). Note that
`N` here refers to the array size (not the iterator length!) which is
expected to be fairly small most of the time.

Unfortunately, this optimization makes the implementation a lot more
involved and requires lots more `unsafe` code. Therefore, quite a few
tests were added to hopefully make sure everything is safe and sound.