Improve function front matter checking
Previously a (invalid) function declared as

    extern "C" const fn test() {}

Would be parsed a invalid block and the error would say to replace
`const` by `{` (or to insert it just before). While this would produce
valid Rust, this was probably not the intent of people writing the whole
`const` keyword in this situation and led to a confusing diagnostic.