Rustdoc accessibility: use real headers for doc items
This is a pretty annoying workaround: the specifications literally
contradict each other on the matter of whether headers may be nested
inside of summaries. We need the headers for people who use outline
mode to browse rustdoc, mostly in screen readers, and the headers
need to be inside the summary tag so that they're visible when the
details tag is collapsed.

HTML5 says it's okay, but the ordinary button semantics in ARIA say
that headers nested in buttons are ignored, and summary is a button.
Which means some screen readers, like NVDA, work, while others,
like JAWS, don't.