bless region-multiple-lifetime-bounds-on-fns-where-clause under polonius
The diagnostics are either the same or better than NLLs: an error that
is seen in the compare-mode, but not visible in the NLL output (probably
because of early return in case of error) is visible in the polonius
output.

A duplicate error is emitted because of a suggestion on the universal
regions.