Generalize `-C prefer-dyanmic` to carry an optional subset of the crates to prefer dynamic linkage on.
Note: carrying an empty set here is *not* the same as `-C prefer-dyanmic=no`
(and the latter is the default behavior for `rustc`). As discussed in detail in
the comments for `enum PreferDynamicSet` (and also in the comments of
`rustc_metadata::dependency_format`), `-C prefer-dynamic=no` asks the compiler
to guess what linkage to use, and it currently guesses all static first, and if
that fails, then it prefers dynamic linkage for all crates that provide both
dynamic and static libraries.

Using an empty set for `-C prefer-dynamic`, on the other hand, means the
compiler should never prefer dynamic linkage over static linkage, even if that
means it won't be able to successfully link the build product.